+++ neatle/src/main/java/si/inova/neatle/operation/CharacteristicSubscriptionImpl.java
@@ -27,7 +27,7 @@ package si.inova.neatle.operation;
 import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.BluetoothDevice;
 import android.content.Context;
-import android.support.annotation.RestrictTo;
+import androidx.annotation.RestrictTo;
 
 import java.util.UUID;