+++ neatle/src/main/java/si/inova/neatle/operation/OperationImpl.java
@@ -31,7 +31,7 @@ import android.bluetooth.BluetoothGattCharacteristic;
 import android.bluetooth.BluetoothGattDescriptor;
 import android.content.Context;
 import android.os.Handler;
-import android.support.annotation.RestrictTo;
+import androidx.annotation.RestrictTo;
 
 import java.util.Collection;
 import java.util.LinkedList;