+++ neatle/src/main/java/si/inova/neatle/util/DeviceManager.java
@@ -28,7 +28,7 @@ import android.annotation.SuppressLint;
 import android.bluetooth.BluetoothDevice;
 import android.bluetooth.BluetoothManager;
 import android.content.Context;
-import android.support.annotation.RestrictTo;
+import androidx.annotation.RestrictTo;
 
 import java.util.HashMap;