+++ sample/src/main/java/si/inova/sample/DeviceDetails.java
@@ -4,12 +4,11 @@ import android.bluetooth.BluetoothDevice;
 import android.bluetooth.BluetoothGattCharacteristic;
 import android.bluetooth.BluetoothGattService;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
-import android.util.Log;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -23,16 +22,11 @@ import si.inova.neatle.ServicesDiscoveredListener;
 import si.inova.neatle.monitor.Connection;
 import si.inova.neatle.monitor.ConnectionMonitor;
 import si.inova.neatle.monitor.ConnectionStateListener;
-import si.inova.neatle.operation.Command;
-import si.inova.neatle.operation.CommandResult;
 import si.inova.neatle.operation.Operation;
-import si.inova.neatle.operation.OperationBuilder;
-import si.inova.neatle.operation.OperationObserver;
 import si.inova.neatle.operation.OperationResults;
 import si.inova.neatle.operation.SimpleOperationObserver;
 import si.inova.neatle.sample.R;
 import si.inova.neatle.sample.ReadAllCommand;
-import si.inova.neatle.sample.fragments.ScannerFragment;
 
 public class DeviceDetails extends AppCompatActivity implements ConnectionStateListener, ServicesDiscoveredListener {
     private BluetoothDevice device;